% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_scores.R
\name{biochemistry_correlation}
\alias{biochemistry_correlation}
\title{Biochemistry Correlation}
\usage{
biochemistry_correlation(query, subject, method = "spearman")
}
\arguments{
\item{query}{Query sequence}

\item{subject}{Subject sequence}

\item{method}{Spearman correlation as default}
}
\value{

}
\description{
Biochemistry Correlation
}
\examples{
query <- 'EVDPIGHLY'
subject <- 'EVDPIGMLY'
biochemistry_correlation(query = query, subject = subject)
}
