% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_utils.R
\name{cross_background}
\alias{cross_background}
\title{Creating immunopeptidomics database}
\usage{
cross_background(off_targets = NULL, allele)
}
\arguments{
\item{off_targets}{A list of off-target candidates. Only 9-mers are supported.}

\item{allele}{Input an MHC Class I allele. Please, check \code{\link{hla_database}}.}
}
\value{
Returns a \code{\link{xrBackground}} object
}
\description{
Peptide database spanning eluted candidates (experimentally validated) and custom (user-defined).
}
\examples{
\dontrun{

# Listing MHC class I alleles
data('hla_database')
View(hla_database)

# Using default immunopeptidomics
background <- cross_background(allele = 'HLA-A*01:01')

# Using MAGE3A off-targets
data('mage_off_targets')

mage_off_targets <- mage_off_targets$peptide_sequence
background <- cross_background(off_targets = mage_off_targets, allele = "HLA-A*01:01")

}
}
