% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_scores.R
\name{cross_bp_summary}
\alias{cross_bp_summary}
\title{Crossdome Summary}
\usage{
cross_bp_summary(query, subject, method = c("pearson", "kendall", "spearman"))
}
\arguments{
\item{query}{Description}

\item{subject}{Description}

\item{method}{Description}
}
\value{

}
\description{
Crossdome Summary
}
\examples{
query <- 'EVDPIGHLY'
subject <- 'EVDPIGMLY'
cross_bp_summary(query = query, subject = subject)
}
