% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_utils.R
\name{cross_browser}
\alias{cross_browser}
\title{Launching Crossdome web application}
\usage{
cross_browser()
}
\description{
Opens an interactive shiny application.
}
\examples{
\dontrun{

 # Opening the shiny application
 cross_browser()

}
}
