% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_compose.R
\name{cross_compose}
\alias{cross_compose}
\title{Crossdome Compose}
\usage{
cross_compose(query, background, allele, position_weight = NULL)
}
\arguments{
\item{query}{Description}

\item{background}{Description}

\item{allele}{Description}

\item{position_weight}{Description}
}
\value{
Description
}
\description{
Crossdome Compose
}
\examples{
data('mage_off_targets')

query <- 'EVDPIGHLY'
off_targets <- mage_off_targets$peptide_sequence
rank <- cross_compose(query = query, background = off_targets)
}
