% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cross_plots.R
\name{cross_enrichment_plot}
\alias{cross_enrichment_plot}
\title{Tissue-specificity barplot}
\usage{
cross_enrichment_plot(object)

\S4method{cross_enrichment_plot}{xrResult}(object)
}
\arguments{
\item{object}{Depends on xrResult object. Run \code{\link{cross_compose}} function.}
}
\value{
Returns a ggplot object
}
\description{
Plot a bar plot summarizing the tissue-specificy groups
}
\examples{
\dontrun{
 cross_enrichment_plot(object = result)
}
}
