% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_utils.R
\name{cross_epitope_properties}
\alias{cross_epitope_properties}
\title{Converting peptide to biochemical properties}
\usage{
cross_epitope_properties(query)
}
\arguments{
\item{query}{A peptide target. Only 9-mers are supported.}
}
\value{
Returns a matrix related to the biochemical profile
}
\description{
Converting peptide to biochemical properties
}
\examples{
cross_epitope_properties('EVDPIGHLY')
}
