% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cross_plots.R
\name{cross_expression_heatmap}
\alias{cross_expression_heatmap}
\title{cross_expression_heatmap}
\usage{
cross_expression_heatmap(object, ...)

\S4method{cross_expression_heatmap}{xrResult}(object, pvalue_threshold = 0.01)
}
\arguments{
\item{object}{Description}

\item{pvalue_threshold}{Description}
}
\value{
Description
}
\description{
cross_expression_heatmap
}
\examples{
if(FALSE) {

result <- cross_expression_matrix(object = result)
cross_expression_heatmap(object = result)
}
}
