% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cross_plots.R
\name{cross_expression_plot}
\alias{cross_expression_plot}
\title{cross_expression_plot}
\usage{
cross_expression_plot(object)

\S4method{cross_expression_plot}{xrResult}(object)
}
\arguments{
\item{object}{Description}

\item{pvalue_threshold}{Description}
}
\value{
Description
}
\description{
cross_expression_plot
}
\examples{
if(FALSE) {

result <- cross_expression_matrix(object = result)
cross_expression_plot(object = result)
}
}
