% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_scores.R
\name{cross_pair_summary}
\alias{cross_pair_summary}
\title{Crossdome relatedness score}
\usage{
cross_pair_summary(query, subject, position_weight = rep(1, 9))
}
\arguments{
\item{query}{A peptide target. Only 9-mers are supported.}

\item{subject}{Putative off-target candidate. Only 9-mers are supported.}

\item{position_weight}{A numeric vector derived from TCR hotspots}
}
\value{
Returns the \code{list} holding Crossdome metrics
}
\description{
Calculates relatedness score between peptides
}
\examples{
query <- 'EVDPIGHLY'
subject <- 'ESDPIVAQY'
cross_pair_summary(query = query, subject = subject)
}
