% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_utils.R
\name{cross_pairwise_plot}
\alias{cross_pairwise_plot}
\title{cross_pairwise_plot}
\usage{
cross_pairwise_plot(query, subject)
}
\arguments{
\item{query}{Description}

\item{subject}{Description}
}
\value{

}
\description{
cross_pairwise_plot
}
\examples{
query <- 'EVDPIGHLY'
subject <- 'EVDPIGMLY'
cross_pairwise_plot(query = query, subject = subject)
}
