% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_utils.R
\name{cross_peptide_properties}
\alias{cross_peptide_properties}
\title{Converts a peptide to biochemical profile}
\usage{
cross_peptide_properties(query)
}
\arguments{
\item{query}{A peptide target. Only 9-mers are supported.}
}
\value{
Returns a matrix related to the biochemical profile.
}
\description{
Converts a peptide to biochemical profile
}
\examples{
cross_peptide_properties('EVDPIGHLY')
}
