% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cross_plots.R
\name{cross_prediction_plot}
\alias{cross_prediction_plot}
\title{Prediction dot plots}
\usage{
cross_prediction_plot(object)

\S4method{cross_prediction_plot}{xrResult}(object)
}
\arguments{
\item{object}{Depends on xrResult object. Run \code{\link{cross_compose}} function.}
}
\value{
Returns a ggplot object
}
\description{
A dot plot showing the immunogenic predictions
}
\examples{
\dontrun{
 cross_prediction_plot(object = result)
}
}
