% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_utils.R
\name{cross_report_table}
\alias{cross_report_table}
\title{cross_report_table}
\usage{
cross_report_table(result)
}
\arguments{
\item{result}{Description}
}
\value{
Description
}
\description{
cross_report_table
}
