% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cross_plots.R
\name{cross_substitution_plot}
\alias{cross_substitution_plot}
\title{Substitution heatmap}
\usage{
cross_substitution_plot(object)

\S4method{cross_substitution_plot}{xrResult}(object)
}
\arguments{
\item{object}{Depends on xrResult object. Run \code{\link{cross_compose}} function.}
}
\value{
Returns a heatmap-like figure showing substitution rate
}
\description{
Plot a heatmap combined with seqlogo displaying amino acid substitutions
}
\examples{
\dontrun{
 cross_substitution_plot(object = result)
}
}
