% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_utils.R
\name{cross_target_expression}
\alias{cross_target_expression}
\title{cross_target_expression}
\usage{
cross_target_expression(epitope)
}
\arguments{
\item{epitope}{Description}
}
\value{
Description
}
\description{
cross_target_expression
}
