% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cross_plots.R
\name{cross_tissues_plot}
\alias{cross_tissues_plot}
\title{Tissue-specificity barplot}
\usage{
cross_tissues_plot(object)

\S4method{cross_tissues_plot}{xrResult}(object)
}
\arguments{
\item{object}{Depends on xrResult object. Run \code{\link{cross_compose}} function.}
}
\value{
Returns a ggplot object
}
\description{
A bar plot summarizing the tissue-specificy groups.
}
\examples{
\dontrun{
 cross_tissues_plot(object = result)
}
}
