% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_utils.R
\name{cross_universe}
\alias{cross_universe}
\title{cross_universe}
\usage{
cross_universe(peptides = NULL, allele)
}
\arguments{
\item{peptides}{Description}

\item{allele}{Description}
}
\value{
Description
}
\description{
cross_universe
}
\examples{
peptides <- mage_off_targets$peptide_sequence
cross_universe(peptides, allele = 'HLA-A*01:01')
}
