% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cross_utils.R
\docType{method}
\name{cross_write}
\alias{cross_write}
\title{cross_write}
\usage{
cross_write(object, ...)

\S4method{cross_write}{xrResult}(
  object,
  file = "",
  append = FALSE,
  quote = FALSE,
  sep = "\\t",
  eol = "\\n",
  na = "NA",
  dec = ".",
  row.names = FALSE,
  col.names = TRUE,
  qmethod = c("escape", "double"),
  fileEncoding = ""
)
}
\arguments{
\item{object}{crossdome xrResult object}

\item{file}{path to output file.}
}
\description{
cross_write
}
