% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtex_database}
\alias{gtex_database}
\title{GTEx Database}
\format{
A data frame with XXXX rows and XXXX variables:
\describe{
\item{ensembl_id}{Description}
\item{genename}{Description}
\item{tissue}{Description}
\item{tpm}{Description}
}
}
\source{
\url{http://}

\url{http://}
}
\usage{
gtex_database
}
\description{
Description
}
\keyword{datasets}
