% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hla_database}
\alias{hla_database}
\title{HLA Database}
\format{
A data frame with 901893 rows and 9 variables:
\describe{
\item{peptide_sequence}{Description}
\item{hla_allele}{Description}
\item{peptide_length}{Description}
\item{resource}{Description}
\item{n_resource}{Description}
}
}
\usage{
hla_database
}
\description{
Description
}
\keyword{datasets}
