% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hpa_database}
\alias{hpa_database}
\title{HPA Expression}
\format{
A data frame with 20090 rows and 42 variables:
\describe{
\item{ensembl_id}{Description}
\item{gene_donor}{Description}
\item{healthy_tissues}{Description}
\item{Group}{Description}
\item{spec_degree}{Description}
\item{tissues}{Description}
}
}
\source{
\url{https://www.proteinatlas.org/}
}
\usage{
hpa_database
}
\description{
Description
}
\keyword{datasets}
