% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mage_off_targets}
\alias{mage_off_targets}
\title{MAGE3A Off-targets (Gee et. al. 2018)}
\format{
A data frame with 61 rows and 5 variables:
\describe{
\item{peptide_sequence}{Description}
\item{hla_allele}{Description}
\item{peptide_length}{Description}
\item{resource}{Description}
\item{n_resource}{Description}
}
}
\source{
\url{https://www.biorxiv.org/content/10.1101/472480v1.full}
}
\usage{
mage_off_targets
}
\description{
Description
Facile method for screening clinical T cell receptors for off-target peptide-HLA reactivity
}
\keyword{datasets}
