% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_scores.R
\name{peptide_alignment_stats}
\alias{peptide_alignment_stats}
\title{Peptide Alignment Stats}
\usage{
peptide_alignment_stats(query, subject)
}
\arguments{
\item{subject}{}
}
\value{

}
\description{
Peptide Alignment Stats
}
\examples{
query <- 'EVDPIGHLY'
subject <- 'EVDPIGMLY'
peptide_alignment_stats(query = query, subject = subject)
}
