% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peptide_annotation}
\alias{peptide_annotation}
\title{Peptide Annotation}
\format{
A data frame with 12158833 rows and 3 variables:
\describe{
\item{peptide_sequence}{Description}
\item{ensembl_id}{Description}
\item{gene_donor}{Description}
}
}
\usage{
peptide_annotation
}
\description{
Description
}
\keyword{datasets}
