% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{xrBackground-class}
\alias{xrBackground-class}
\title{Class "xrBackground"
Description}
\description{
Class "xrBackground"
Description
}
\section{Slots}{

\describe{
\item{\code{peptides}}{Description}

\item{\code{allele}}{Description}
}}

\keyword{classes}
