% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_utils.R
\name{cross_browser}
\alias{cross_browser}
\title{cross_epitope_properties}
\usage{
cross_browser()
}
\arguments{
\item{epitope}{Description}
}
\value{

}
\description{
cross_epitope_properties
}
\examples{
cross_epitope_properties('EVDPIGHLY')
}
