% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_compose.R
\name{cross_compose}
\alias{cross_compose}
\title{Title}
\usage{
cross_compose(query, universe = HLA_LIST, tcr_weights, widget = TRUE)
}
\arguments{
\item{query}{Description}

\item{tcr_weights}{Description}

\item{subject}{Description}

\item{hla_allele}{Description}
}
\value{

}
\description{
Title
}
\examples{

}
