% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cross_compose.R
\docType{methods}
\name{cross_compose}
\alias{cross_compose}
\title{cross_compose}
\usage{
cross_compose(query, object, position_weight = rep(1, 9))

\S4method{cross_compose}{ANY,xrBackground}(query, object, position_weight = rep(1, 9))
}
\arguments{
\item{query}{Description}

\item{position_weight}{Description}

\item{background}{Description}
}
\value{
Description
}
\description{
cross_compose
}
\examples{
data('mage_off_targets')

mage_off_targets <- mage_off_targets$peptide_sequence
background <- cross_universe(off_targets = mage_off_targets, allele = "HLA-A*01:01")
result <- cross_compose(query = query, background = background)
}
