% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_compose.R
\name{cross_compose}
\alias{cross_compose}
\title{Crossdome Compose}
\usage{
cross_compose(query, subject, widgets = FALSE)
}
\arguments{
\item{query}{Description}

\item{subject}{Description}

\item{widgets}{Description}
}
\value{

}
\description{
Crossdome Compose
}
\examples{
data('mage_off_targets')

query <- 'EVDPIGHLY'
subject <- mage_off_targets$peptide_sequence
rank <- cross_compose(query = query, subject = subject, widgets = FALSE)
}
