% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cross_compose.R
\docType{methods}
\name{cross_compose}
\alias{cross_compose}
\title{Performing Crossdome screening}
\usage{
cross_compose(query, background, position_weight = rep(1, 9))

\S4method{cross_compose}{ANY,xrBackground}(query, background, position_weight = rep(1, 9))
}
\arguments{
\item{query}{A peptide target. Only 9-mers are supported.}

\item{background}{Background object create with \code{\link{cross_universe}}}

\item{position_weight}{A numeric vector derived from TCR hotspots}
}
\value{
Returns a \code{\link{xrResult}} object containing relatedness ranking
}
\description{
Predicts relatedness among peptides in a given database. Low values are associated with cross-reactive candidates.
}
\examples{
\dontrun{
data('mage_off_targets')

mage_off_targets <- mage_off_targets$peptide_sequence
background <- cross_universe(off_targets = mage_off_targets, allele = "HLA-A*01:01")
result <- cross_compose(query = query, background = background

}
}
