% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_plots.R
\name{cross_expression_heatmap}
\alias{cross_expression_heatmap}
\title{cross_expression_heatmap}
\usage{
cross_expression_heatmap(result, epitope, top = 50)
}
\arguments{
\item{result}{Description}

\item{epitope}{Description}

\item{top}{Description}
}
\value{
Description
}
\description{
cross_expression_heatmap
}
\examples{
epitope <- c("EVDPIGHLY", "ESDPIVAQY")
cross_expression_heatmap(epitope = epitope)
}
