% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cross_utils.R
\docType{methods}
\name{cross_expression_matrix}
\alias{cross_expression_matrix}
\title{cross_expression_matrix}
\usage{
cross_expression_matrix(object)

\S4method{cross_expression_matrix}{xrResult}(object)
}
\arguments{
\item{object}{Description}
}
\value{
Description
}
\description{
cross_expression_matrix
}
\examples{

result <- crossdome::mage_result
cross_expression_matrix(object = result)
}
