% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cross_utils.R
\docType{methods}
\name{cross_expression_matrix}
\alias{cross_expression_matrix}
\title{Collecting gene-donor expression data}
\usage{
\S4method{cross_expression_matrix}{xrResult}(object, rank = NULL, pvalue_threshold = 0.01)
}
\arguments{
\item{object}{Depends on xrResult object. Run \code{\link{cross_compose}} function.}

\item{rank}{A numeric value to filter number of candidates}

\item{pvalue_threshold}{P-value threshold}
}
\value{
Return a matrix containing mRNA expression across healthy tissues.
}
\description{
Extracts gene donor mRNA expression based on CR candidates.
}
\examples{
\dontrun{
result <- crossdome::mage_result
cross_expression_matrix(object = result)
}
}
