% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cross_plots.R
\name{cross_expression_plot}
\alias{cross_expression_plot}
\title{Expression heatmap}
\usage{
cross_expression_plot(object)

\S4method{cross_expression_plot}{xrResult}(object)
}
\arguments{
\item{object}{Depends on xrResult object. Run \code{\link{cross_compose}} function.}
}
\value{
Returns a combined \strong{ComplexHeatmap} object
}
\description{
A heatmap presenting the gene donor expression profile.
}
\examples{
\dontrun{

result <- cross_expression_matrix(object = result)
cross_expression_plot(object = result)
}
}
