% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_scores.R
\name{cross_pair_summary}
\alias{cross_pair_summary}
\title{Crossdome Summary}
\usage{
cross_pair_summary(
  query,
  subject,
  position_weight = NULL,
  method = c("pearson", "kendall", "spearman")
)
}
\arguments{
\item{query}{Description}

\item{subject}{Description}

\item{position_weight}{Description}

\item{method}{Description}
}
\value{
Description
}
\description{
Crossdome Summary
}
\examples{
query <- 'EVDPIGHLY'
subject <- 'ESDPIVAQY'
cross_pair_summary(query = query, subject = subject)
}
