% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_plots.R
\name{cross_pairwise_plot}
\alias{cross_pairwise_plot}
\title{Correlation plot}
\usage{
cross_pairwise_plot(query, subject)
}
\arguments{
\item{query}{Peptide target. Only 9-mers are supported.}

\item{subject}{Putative off-target candidate. Only 9-mers are supported.}
}
\value{
Returns a \strong{corrplot} object
}
\description{
Correlation plot based on two peptide / biochemical profiles.
}
\examples{
query <- 'EVDPIGHLY'
subject <- 'ESDPIVAQY'
cross_pairwise_plot(query = query, subject = subject)
}
