% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cross_plots.R
\name{cross_substitution_matrix}
\alias{cross_substitution_matrix}
\title{cross_substitution_matrix}
\usage{
cross_substitution_matrix(object, ...)

\S4method{cross_substitution_matrix}{xrResult}(object, pvalue_threshold = 0.01)
}
\arguments{
\item{object}{Description}

\item{top}{Description}
}
\value{
Description
}
\description{
cross_substitution_matrix
}
\examples{
if(FALSE) {
 cross_substitution_matrix(object = result)
}
}
