% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cross_plots.R
\name{cross_substitution_plot}
\alias{cross_substitution_plot}
\title{cross_substitution_plot}
\usage{
cross_substitution_plot(object)

\S4method{cross_substitution_plot}{xrResult}(object)
}
\arguments{
\item{object}{Description}

\item{top}{Description}
}
\value{
Description
}
\description{
cross_substitution_plot
}
\examples{
if(FALSE) {
 cross_substitution_plot(object = result)
}
}
