% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_utils.R
\name{cross_universe}
\alias{cross_universe}
\title{cross_universe}
\usage{
cross_universe(off_targets = NULL, allele)
}
\arguments{
\item{allele}{Description}

\item{peptides}{Description}
}
\value{
Description
}
\description{
cross_universe
}
\examples{
# Using default immunopeptidomics
background <- cross_universe(allele = 'HLA-A*01:01')

# Using MAGE3A off-targets
data('mage_off_targets')

mage_off_targets <- mage_off_targets$peptide_sequence
background <- cross_universe(off_targets = mage_off_targets, allele = "HLA-A*01:01")
}
