% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cross_utils.R
\name{cross_write}
\alias{cross_write}
\title{Data Output}
\usage{
cross_write(object, ...)

\S4method{cross_write}{xrResult}(
  object,
  file = "",
  append = FALSE,
  quote = FALSE,
  sep = "\\t",
  eol = "\\n",
  na = "NA",
  dec = ".",
  row.names = FALSE,
  col.names = TRUE,
  qmethod = c("escape", "double"),
  fileEncoding = ""
)
}
\arguments{
\item{object}{Depends on xrResult object. Run \code{\link{cross_compose}} function.}

\item{file}{File or connection to write to.}
}
\description{
Export Crossdome object to a tsv file
}
