% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peptide_annotation}
\alias{peptide_annotation}
\title{Peptide Annotation}
\format{
A data frame with XXXX rows and XXXX variables:
\describe{
\item{peptide_sequence}{Description}
\item{ensembl_id}{Description}
\item{gene_donor}{Description}
\item{n_gene}{Description}
}

A data frame with XXXX rows and XXXX variables:
\describe{
\item{peptide_sequence}{Description}
\item{ensembl_id}{Description}
\item{gene_donor}{Description}
\item{n_gene}{Description}
}
}
\source{
\url{http://}

\url{http://}

\url{http://}

\url{http://}
}
\usage{
peptide_annotation

peptide_annotation
}
\description{
Description

Description
}
\keyword{datasets}
