% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{xrBackground-class}
\alias{xrBackground-class}
\title{Class "xrBackground"}
\description{
An object holding data from \code{\link{hla_database}}
}
\section{Slots}{

\describe{
\item{\code{allele}}{Description}

\item{\code{peptides}}{Description}

\item{\code{stats}}{A slot containing statistics}
}}

\keyword{classes}
