% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{xrResult-class}
\alias{xrResult-class}
\title{Class "xrResult"}
\description{
The Crossdome main object
}
\section{Slots}{

\describe{
\item{\code{query}}{A peptide target. Only 9-mers are supported.}

\item{\code{result}}{Crossdome ranking data.frame}

\item{\code{allele}}{Input a MHC Class I allele}

\item{\code{expression}}{Expression data slot}

\item{\code{analysis}}{Sequence and immunogenicity slot}

\item{\code{position_weight}}{A numeric vector derived from TCR hotspots}

\item{\code{timestamp}}{Execution timestamp}
}}

\keyword{classes}
