% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{xrResult-class}
\alias{xrResult-class}
\title{Class "xrResult"
Description}
\description{
Class "xrResult"
Description
}
\section{Slots}{

\describe{
\item{\code{result}}{Description}

\item{\code{allele}}{Allele}

\item{\code{position_weight}}{numeric}
}}

\keyword{classes}
